test: db-remove: Verify that it accepts pkgname (in addition to pkgbase)
It is important that db-remove be able to remove a single pkgname, without
being able to look it up by pkgbase in SVN.  For instance, when a split
package update removes one of its members; there will be no reference to
the removed pkgname in SVN, and it won't be removed by db-update.  If
db-remove doesn't accept pkgnames, then this outdated orphan could not be
removed.